--- a/server/service/src/report/convert_to_excel.rs
+++ b/server/service/src/report/convert_to_excel.rs
@@ -192,7 +192,7 @@ fn apply_data_rows(
 
     for row in body.rows_and_cells().into_iter() {
         // Insert new row below (leave any footer from the excel template in place)
-        sheet.insert_new_row(&(row_idx + 1), &1);
+        // sheet.insert_new_row(&(row_idx + 1), &1);
 
         // Duplicate any formulae/formatting to the next row before populating
         for col in 0..sheet.get_highest_column() {