<!-- IMPORTANT!
  - Every PR must reference an issue; this helps to explain the intent of the PR
 -->

Fixes #8415 

# 👩🏻‍💻 What does this PR do?

Before:
100 rows: 0.12s
1000 rows: 9.5s
10000 rows: 1123s 
![flamegraph_before](https://github.com/user-attachments/assets/a56fdb08-a5d3-4b2e-b8a1-1bc938080b1b)

It appears the vast majority of time is `Worksheet::insert_new_row()`. Perhaps quadratic if not exponential big O as for every row added I am guessing it executes `rebuild_map_and_indicies`, as the sheet grows this gets slower. Oof. Soooo I removed it:

After: 10000 rows 1.6s
![flamegraph](https://github.com/user-attachments/assets/805b0929-baf6-4d0a-8829-a9032d494cfc)

(note our `apply_data_rows` is now a very minor player, and most of the time is in `scraper` code parsing the html rather than `umya` building the spreadsheet)

This appears to work correctly for item usage report which just does default direct mapping of html to excel, however I expect the current change to break footers if using a xlsx template. The intent with the insert rows is to push the footer rows of the template down. 

Turns out you can just `insert_new_row` with second parameter being how many rows to add, so you only really need to call it once. So now:

10000 rows: 1.6s (crudely timed maybe 50ms slower than without adding rows at all)
![flamegraph](https://github.com/user-attachments/assets/e4048360-c752-4453-8260-800ab4e8f296)

Virtually identical to not adding rows at all, but respects the excel template footer by pushing it down.

## 💌 Any notes for the reviewer?

So ya like flamegraphs? I spent the morning figuring out how to get a cli tool working on a particular test so you don't have to:

```sh
cargo install flamegraph
```
https://github.com/flamegraph-rs/flamegraph

```sh
cargo flamegraph --dev --unit-test service -- report::convert_to_excel::report_to_excel_test::test_generate_excel_perf
```

(Note, I'm using --dev cause otherwise it compiles release which takes forever. If you were flamegraphing the actual OMS server while running you might benefit from running release binary.)

# 🧪 Testing

<!-- Explain the steps you'd take to test the changes of this PR manually -->

- [ ] OMS site with 4000+ items
- [ ] item usage report takes <1min (definitely not 5mins+)

# 📃 Documentation

- [ ] **Part of an epic**: documentation will be completed for the feature as a whole
- [x] **No documentation required**: no user facing changes or a bug fix which isn't a change in behaviour
- [ ] **These areas should be updated or checked**: <!-- _(e.g.)_ New `issued` column in `Requisitions` indicates stock quantity already in shipments -->
  1.
  3.


# 📃 Reviewer Checklist

The PR Reviewer(s) should fill out this section before approving the PR

**Breaking Changes**
- [x] No Breaking Changes in the Graphql API
- [ ] Technically some Breaking Changes but not expected to impact any integrations

**Issue Review**
- [x] All requirements in original issue have been covered
- [ ] A follow up issue(s) have been created to cover additional requirements

**Tests Pass**
- [x] Postgres
- [x] SQLite
- [x] Frontend