@Chris-Petty 2025-07-17
@lache-melvin it is ready! If the comments didn't give it away, the new test for the excel export with template was 90% vibe coded and then fixed haha. The tokio timeout is artisanally crafted on the other hand.